Renaming _prettyprint to prettyprint (no longer private)
