+++ src/awkward/highlevel.py
@@ -31,7 +31,7 @@ from awkward._pickle import (
     unpickle_array_schema_1,
     unpickle_record_schema_1,
 )
-from awkward._prettyprint import Formatter
+from awkward.prettyprint import Formatter
 from awkward._regularize import is_non_string_like_iterable
 from awkward._typing import Any, TypeVar
 from awkward._util import STDOUT
@@ -1291,15 +1291,15 @@ class Array(NDArrayOperatorsMixin, Iterable, Sized):
         )
 
     def __str__(self):
-        import awkward._prettyprint
+        import awkward.prettyprint
 
-        return awkward._prettyprint.valuestr(self, 1, 80)
+        return awkward.prettyprint.valuestr(self, 1, 80)
 
     def __repr__(self):
         return self._repr(80)
 
     def _repr(self, limit_cols):
-        import awkward._prettyprint
+        import awkward.prettyprint
 
         try:
             pytype = super().__getattribute__("__name__")
@@ -1322,7 +1322,7 @@ class Array(NDArrayOperatorsMixin, Iterable, Sized):
                     limit_cols - len(pytype) - len(" type='...'") - 3,
                 ),
             )
-        valuestr = valuestr + " " + awkward._prettyprint.valuestr(self, 1, strwidth)
+        valuestr = valuestr + " " + awkward.prettyprint.valuestr(self, 1, strwidth)
 
         length = max(3, limit_cols - len(pytype) - len("type='...'") - len(valuestr))
         if len(typestr) > length:
@@ -1363,11 +1363,11 @@ class Array(NDArrayOperatorsMixin, Iterable, Sized):
         key is ignored; instead, a `"bytes"` and/or `"str"` key is considered when formatting
         string values, falling back upon `"str_kind"`.
         """
-        import awkward._prettyprint
+        import awkward.prettyprint
 
         formatter_impl = Formatter(formatter, precision=precision)
 
-        valuestr = awkward._prettyprint.valuestr(
+        valuestr = awkward.prettyprint.valuestr(
             self, limit_rows, limit_cols, formatter=formatter_impl
         )
         if type:
@@ -2161,15 +2161,15 @@ class Record(NDArrayOperatorsMixin):
         )
 
     def __str__(self):
-        import awkward._prettyprint
+        import awkward.prettyprint
 
-        return awkward._prettyprint.valuestr(self, 1, 80)
+        return awkward.prettyprint.valuestr(self, 1, 80)
 
     def __repr__(self):
         return self._repr(80)
 
     def _repr(self, limit_cols):
-        import awkward._prettyprint
+        import awkward.prettyprint
 
         pytype = type(self).__name__
 
@@ -2189,7 +2189,7 @@ class Record(NDArrayOperatorsMixin):
                     limit_cols - len(pytype) - len(" type='...'") - 3,
                 ),
             )
-        valuestr = valuestr + " " + awkward._prettyprint.valuestr(self, 1, strwidth)
+        valuestr = valuestr + " " + awkward.prettyprint.valuestr(self, 1, strwidth)
 
         length = max(3, limit_cols - len(pytype) - len("type='...'") - len(valuestr))
         if len(typestr) > length:
@@ -2229,10 +2229,10 @@ class Record(NDArrayOperatorsMixin):
         key is ignored; instead, a `"bytes"` and/or `"str"` key is considered when formatting
         string values, falling back upon `"str_kind"`.
         """
-        import awkward._prettyprint
+        import awkward.prettyprint
 
         formatter_impl = Formatter(formatter, precision=precision)
-        valuestr = awkward._prettyprint.valuestr(
+        valuestr = awkward.prettyprint.valuestr(
             self, limit_rows, limit_cols, formatter=formatter_impl
         )
         if type:
